apps/api/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abase-api"
version = "0.1.0"
description = "ABASE Manager API service"
authors = [{ name = "ABASE" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.29",
    "psycopg[binary]==3.2.1",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "python-jose==3.3.0",
    "httpx==0.27.0",
    "redis==5.0.4",
    "fastapi-sse==0.2.1",
    "aiofiles==23.2.1",
    "python-multipart==0.0.9",
    "pydantic==2.6.4",
    "pydantic-settings==2.2.1",
    "passlib[bcrypt]==1.7.4",
    "uvicorn-worker==0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.1.1",
    "pytest-asyncio==0.23.6",
    "httpx==0.27.0",
    "ruff==0.3.5",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
python_files = ["tests.py", "test_*.py"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
